Fix in collision and health system.
Now in the collision system, if a collision with a -> b happens, a
collision b -> a will be ignored during the same update (can happen if both entities are moving towards each
other).

A basic health system has been added that can remove health from entities and decide if they are alive or not.