Update the animation class
Now each frame in an animation has it own time.
This enables more kinds of animations.

Add a new sprite sheet for "CharacterBandana"
with an added death animation.

Add so that player entities doesn't get destroyed
when killed. The entity will still exist and the last
animation frame will be shown where the player died.